feat(email): enhance email templates and restructure auth email components
